removed unused code from the engine
